Moved examples to their own module
