+++ pysimpleapp/examples/example_tm.py
@@ -3,8 +3,8 @@ import threading
 from queue import Queue
 
 from pysimpleapp.message import Message
-from thread_manager import ThreadManager
-from examples import ExampleSingleRunThread
+from pysimpleapp.thread_manager import ThreadManager
+from pysimpleapp.examples.thread_examples import ExampleSingleRunThread
 
 if __name__ == "__main__":
